Development.GitRev.Typed.OsString

Description

Development.GitRev.Typed for OsString.

Since: 2.0

Basic functions

These functions are simple, merely a typed version of Development.GitRev's API.

NOTE: These functions do not error if git fails to run, opting instead to return some default value (e.g. string UNKNOWN, boolean False).

gitBranch :: Code Q OsString

Return the branch (or tag) name of the current git commit, or UNKNOWN if not in a git repository. For detached heads, this will just be HEAD.

Examples

Expand

λ. $$(gitBranch)
"main"

>>> $$(gitBranch)
...

Since: 2.0

gitCommitCount :: Code Q OsString

Return the number of commits in the current head.

Examples

Expand

λ. $$(gitCommitCount)
"47"

>>> $$(gitCommitCount)
...

Since: 2.0

gitCommitDate :: Code Q OsString

Return the commit date of the current head.

Examples

Expand

λ. $$(gitCommitDate)
"Mon Apr 14 22:14:44 2025 +1200"

>>> $$(gitCommitDate)
...

Since: 2.0

gitDescribe :: Code Q OsString

Return the long git description for the current git commit, or UNKNOWN if not in a git repository.

Examples

Expand

λ. $$(gitDescribe)
"e67e943"

>>> $$(gitDescribe)
...

Since: 2.0

gitDiff :: Code Q OsString

Return the diff of the working copy with HEAD.

Examples

Expand

λ. $$(gitDiff)
"Mon Apr 14 22:14:44 2025 +1200"

>>> $$(gitDiff)
...

gitDirty :: Code Q Bool

Return True if there are non-committed files present in the repository.

Examples

Expand

λ. $$(gitDirty)
False

>>> $$(gitDirty)
...

Since: 2.0

gitDirtyTracked :: Code Q Bool

Return True if there are non-commited changes to tracked files present in the repository.

Examples

Expand

λ. $$(gitDirtyTracked)
False

>>> $$(gitDirtyTracked)
...

Since: 2.0

gitHash :: Code Q OsString

Return the hash of the current git commit, or UNKNOWN if not in a git repository.

Examples

Expand

λ. $$(gitHash)
"e67e943dd03744d3f93c21f84e127744e6a04543"

>>> $$(gitHash)
...

Since: 2.0

gitShortHash :: Code Q OsString

Return the short hash of the current git commit, or UNKNOWN if not in a git repository.

Examples

Expand

λ. $$(gitShortHash)
"e67e943"

>>> $$(gitShortHash)
...

Since: 2.0

gitTree :: Code Q OsString

Return the hash of the current tree.

Examples

Expand

λ. $$(gitTree)
"Mon Apr 14 22:14:44 2025 +1200"

>>> $$(gitTree)
...

Custom behavior

These functions allow defining custom behavior. For instance, using the primitive getHashQ and combinator projectError, we can define a variant of gitHash that instead fails to compile if there are any problems with git:

>>> :{
  let gitHashOrDie :: Code Q OsString
      gitHashOrDie = qToCode $ projectError gitHashQ
:}

We can also define a function that falls back to an environment variable, in case the git command fails. firstSuccessQ takes the first action that returns Right.

>>> :{
  let gitHashEnv :: OsString -> Code Q (Either (Errors GitOrEnvLookupError) OsString)
      gitHashEnv var =
        qToCode $
          firstSuccessQ
            (embedGitError gitHashQ)
            [embedEnvLookupError $ envValQ var]
:}

Naturally, these can be combined:

>>> :{
  let gitHashEnvOrDie :: OsString -> Code Q OsString
      gitHashEnvOrDie var =
        qToCode
          . projectError
          $ firstSuccessQ
            (embedGitError gitHashQ)
            [embedEnvLookupError $ envValQ var]
:}

"Out-of-tree" builds

Custom definitions are particularly useful for "out-of-tree" builds, where the build takes place outside of the normal git tree.

These builds present a problem, as we normally rely on building in the project directory where the .git directory is easy to locate. For example, while gitHash will work for 'cabal build', it will not work for 'nix build' or 'cabal install'. Fortunately, there are workarounds, both relying on passing the right data via environment variables.

Passing the git directory.

For situations where we can pass the current directory during installation e.g.

$ export EXAMPLE_HOME=$(pwd); cabal install example

We can define

>>> :{
  let gitHashSrcDir :: Code Q OsString
      gitHashSrcDir =
        qToCode
        . projectStringUnknown
        $ firstSuccessQ
          -- 1. We first try normal gitHashQ.
          (embedGitError gitHashQ)
          -- 2. If that fails, we try again in the directory pointed
          --    to by "EXAMPLE_HOME".
          [runGitInEnvDirQ [osstr|EXAMPLE_HOME|] gitHashQ]
:}

If the initial call to gitHashQ fails, then we will try again, running the command from the directory pointed to by EXAMPLE_HOME. firstSuccessQ ensures we do not run the second action unless the first fails.

Passing the value itself.

This approach can work well with nix, as nix flakes provides a variety of revisions via its self interface. For example:

  # Injecting the git hash via EXAMPLE_HASH where drv is the normal
  # derivation.
  drv.overrideAttrs (oldAttrs: {
    # Also: self.shortRev, self.dirtyShortRev
    EXAMPLE_HASH = "${self.rev or self.dirtyRev}";
  });

Then we can define

>>> :{
  let gitHashVal :: Code Q OsString
      gitHashVal =
        qToCode
        . projectStringUnknown
        $ firstSuccessQ
          -- 1. We first try normal gitHashQ.
          (embedGitError gitHashQ)
          -- 2. If that fails, get the value directly from
          --    "EXAMPLE_HASH".
          [embedEnvLookupError $ envValQ [osstr|EXAMPLE_HASH|]]
:}

Once again, if the first attempt fails, we will run the second action, looking for the value of EXAMPLE_HASH.

Finally, we can compose these together to make a function that works for all three cases:

>>> :{
  let gitHashValSrc :: Code Q OsString
      gitHashValSrc =
        qToCode
          . projectStringUnknown
          $ firstSuccessQ
            (embedGitError gitHashQ)
            [ embedEnvLookupError $ envValQ [osstr|EXAMPLE_HASH|],
              runGitInEnvDirQ [osstr|EXAMPLE_HOME|] gitHashQ
            ]
:}

A final note on laziness: As alluded to above, Q's default semigroup instance is not lazy enough:

>>> :{
  $$( qToCode $
      (runIO (putStrLn "in q1") $> (Right "q1") :: Q (Either () String))
        <> (runIO (putStrLn "in q2") $> Left ())
    )
:}
in q1
in q2
Right "q1"

For this reason, we introduce the QFirst newtype:

>>> :{
  $$( qToCode $ unQFirst $
      (mkQFirst $ runIO (putStrLn "in q1") $> (Right "q1") :: QFirst () String)
        <> (mkQFirst $ runIO (putStrLn "in q2") $> Left ())
    )
:}
in q1
Right "q1"

The convenience function

  firstSuccessQ :: Q (Either e a) -> [Q (Either e a)] -> Q (Either (Errors e) a)

utilizes QFirst for sequencing a series of Q actions, stopping after the first success.

Git Primitives

gitBranchQ :: Q (Either GitError OsString)

Returns the current git branch.

Examples

Expand

>>> $$(qToCode gitBranchQ)
Right ...

Since: gitrev-internal-0.1

gitCommitCountQ :: Q (Either GitError OsString)

Returns the git commit count.

Examples

Expand

>>> $$(qToCode gitCommitCountQ)
Right ...

Since: gitrev-internal-0.1

gitCommitDateQ :: Q (Either GitError OsString)

Returns the latest git commit date.

Examples

Expand

>>> $$(qToCode gitCommitDateQ)
Right ...

Since: gitrev-internal-0.1

gitDescribeQ :: Q (Either GitError OsString)

Returns the git description.

Examples

Expand

>>> $$(qToCode gitDescribeQ)
Right ...

Since: gitrev-internal-0.1

gitDiffQ :: Q (Either GitError OsString)

Return the diff of the working copy with HEAD.

Examples

Expand

>>> $$(qToCode gitDiffQ)
Right ...

Since: gitrev-internal-0.1

gitDirtyQ :: Q (Either GitError Bool)

Returns the git dirty status.

Examples

Expand

>>> $$(qToCode gitDirtyQ)
Right ...

Since: gitrev-internal-0.1

gitDirtyTrackedQ :: Q (Either GitError Bool)

Returns the git dirty status, ignoring untracked files.

Examples

Expand

>>> $$(qToCode gitDirtyTrackedQ)
Right ...

Since: gitrev-internal-0.1

gitHashQ :: Q (Either GitError OsString)

Returns the latest git hash.

Examples

Expand

>>> $$(qToCode gitHashQ)
Right ...

Since: gitrev-internal-0.1

gitShortHashQ :: Q (Either GitError OsString)

Returns the latest git short hash.

Examples

Expand

>>> $$(qToCode gitShortHashQ)
Right ...

Since: gitrev-internal-0.1

gitTreeQ :: Q (Either GitError OsString)

Returns the hash of the current tree.

Examples

Expand

>>> $$(qToCode gitTreeQ)
Right ...

Since: gitrev-internal-0.1

Running your own git actions

runGit

Arguments

:: [OsString]
Arguments to git.

-> IndexUsed
Whether the index is used.

-> Q (Either GitError OsString)

Runs git with the arguments. If IdxUsed is passed, it is tracked for recompliation purposes.

Since: gitrev-internal-0.1

data IndexUsed

Type to flag if the git index is used or not in a call to runGit.

Since: gitrev-internal-0.1

Constructors

IdxUsed
The git index is used.

Since: gitrev-internal-0.1

IdxNotUsed
The git index is not used.

Since: gitrev-internal-0.1

Environment lookup

envValQ

Arguments

:: OsString
The environment variable k.

-> Q (Either EnvLookupError OsString)
The result v or an error.

Performs an environment variable lookup in Q.

Examples

Expand

>>> setEnv "SOME_VAR" "val"
>>> $$(qToCode $ envValQ [osstr|SOME_VAR|])
Right "val"

Since: gitrev-internal-0.1

runGitInEnvDirQ

Arguments

:: OsString
Environment variable pointing to a directory path, in which we run the git process.

-> Q (Either GitError a)
Git process to run.

-> Q (Either GitOrEnvLookupError a)
The result.

runGitInEnvDirQ var q runs q in the directory given by the environment variable.

Examples

Expand

>>> setEnv "SOME_DIR" "./"
>>> $$(qToCode $ runGitInEnvDirQ [osstr|SOME_DIR|] gitHashQ)
Right ...

Since: gitrev-internal-0.1

Q to Code

qToCode :: Lift a => Q a -> Code Q a

Lifts a Q computation to Code, for usage with typed TH.

Since: 2.0

Q Combinators

First success

data QFirst e a

Wrapper for Q over Either with a lazier Semigroup. With this, we can run:

  mkQFirst q1 <> mkQFirst q2

This will only execute q2 if q1 returns Left, unlike Q's normal Semigroup instance.

QFirst also collects all errors in Errors.

Since: gitrev-internal-0.1

mkQFirst :: Q (Either e a) -> QFirst e a

Wraps a Q computation in QFirst.

Since: gitrev-internal-0.1

unQFirst :: QFirst e a -> Q (Either (Errors e) a)

Unwraps QFirst.

Since: gitrev-internal-0.1

firstSuccessQ :: Q (Either e a) -> [Q (Either e a)] -> Q (Either (Errors e) a)

firstSuccessQ q qs takes the first qi in q : qs that returns Right, without executing any qj for j > i. If there are no Right's, returns all errors.

Examples

Expand

>>> :{
   $$( qToCode $
         firstSuccessQ
           (pure (Left GitNotFound))
           [ gitHashQ,
             error "oh no"
           ]
     )
:}
Right ...

Since: gitrev-internal-0.1

data Errors e

Collects multiple errors. Intended for displaying multiple exceptions via displayException.

Since: gitrev-internal-0.1

mkErrors :: NonEmpty e -> Errors e

Wraps a type in Errors.

Since: gitrev-internal-0.1

unErrors :: Errors e -> NonEmpty e

Unwraps Errors.

Since: gitrev-internal-0.1

Eliminating Either

projectStringUnknown :: Functor f => f (Either e OsString) -> f OsString

Projects Left to the string UNKNOWN.

Examples

Expand

>>> :{
  let gitHashUnknownQ :: Q OsString
      gitHashUnknownQ = projectStringUnknown gitHashQ
  -- inling gitHashUnknownQ here due to stage restriction
  in $$(qToCode $ projectStringUnknown gitHashQ)
:}
...

>>> $$(qToCode $ projectStringUnknown (pure $ Left ()))
"UNKNOWN"

Since: gitrev-internal-0.1

projectConst :: forall f e a. Functor f => a -> f (Either e a) -> f a

Projects Left to the given value.

Examples

Expand

>>> $$(qToCode $ projectConst "FAILURE" (pure $ Left ()))
"FAILURE"

Since: gitrev-internal-0.1

projectFalse :: Functor f => f (Either e Bool) -> f Bool

Projects Left to False.

Examples

Expand

>>> $$(qToCode $ projectFalse (pure $ Left ()))
False

Since: gitrev-internal-0.1

projectError :: forall f e a. (Exception e, Functor f) => f (Either e a) -> f a

Projects Left via error, rendering via displayException. Hence an error will cause a compilation failure.

Examples

Expand

>>> :{
  let gitHashOrDieQ :: Q Bool
      gitHashOrDieQ = projectError gitDirtyQ
  in $$(qToCode $ projectError gitDirtyQ)
:}
...

Since: gitrev-internal-0.1

projectErrorMap :: Functor f => (e -> String) -> f (Either e a) -> f a

Projects Left via error, rendering via the given function. Hence an error will cause a compilation failure.

Examples

Expand

>>> :{
  let gitHashOrDieQ :: Q Bool
      gitHashOrDieQ = (projectErrorMap show) gitDirtyQ
  in $$(qToCode $ (projectErrorMap show) gitDirtyQ)
:}
...

Since: gitrev-internal-0.1

projectLeft :: Functor f => (e -> a) -> f (Either e a) -> f a

Projects Left via the given function.

Since: gitrev-internal-0.1

Errors

data GitOrEnvLookupError

Git or env lookup error.

Since: gitrev-internal-0.1

Constructors

GitOrEnvLookupGit GitError
Since: gitrev-internal-0.1

GitOrEnvLookupEnvLookup EnvLookupError
Since: gitrev-internal-0.1

data GitError

Errors that can be encountered with git.

Since: gitrev-internal-0.1

Constructors

GitNotFound
Since: gitrev-internal-0.1

GitRunError OsString
Since: gitrev-internal-0.1

newtype EnvLookupError

Environment variable lookup failure. The value is the variable we attempted to look up.

Since: gitrev-internal-0.1

Constructors

MkEnvLookupError OsString

Utilities

embedGitError

Arguments

:: forall f p a. (Bifunctor p, Functor f)

=> f (p GitError a)
.

-> f (p GitOrEnvLookupError a)

Embeds a GitError in the larger GitOrEnvLookupError.

Examples

Expand

>>> :{
  let q :: Q (Either GitError ())
      q = pure (Left GitNotFound)
  in runQ $ embedGitError q
:}
Left (GitOrEnvLookupGit GitNotFound)

Since: gitrev-internal-0.1

embedEnvLookupError

Arguments

:: forall f p a. (Bifunctor p, Functor f)

=> f (p EnvLookupError a)
.

-> f (p GitOrEnvLookupError a)

Embeds a EnvLookupError in the larger GitOrEnvLookupError.

Examples

Expand

>>> :{
  let q :: Q (Either EnvLookupError ())
      q = pure (Left $ MkEnvLookupError [osstr|VAR|])
  in runQ $ embedEnvLookupError q
:}
Left (GitOrEnvLookupEnvLookup (MkEnvLookupError "VAR"))

Since: gitrev-internal-0.1

joinEnvLookupGitErrors

Arguments

:: (Bifunctor p, forall e. Monad (p e))

=> p EnvLookupError (p GitError a)
.

-> p GitOrEnvLookupError a

Utility function for joining lookup and git errors.

Examples

Expand

>>> :{
  let e :: Either EnvLookupError (Either GitError ())
      e = Right (Left GitNotFound)
  in joinEnvLookupGitErrors e
:}
Left (GitOrEnvLookupGit GitNotFound)

Since: gitrev-internal-0.1

joinGitEnvLookupErrors

Arguments

:: (Bifunctor p, forall e. Monad (p e))

=> p GitError (p EnvLookupError a)
.

-> p GitOrEnvLookupError a

Utility function for joining git and lookup errors.

Examples

Expand

>>> :{
  let e :: Either GitError (Either EnvLookupError ())
      e = Right (Left $ MkEnvLookupError [osstr|VAR|])
  in joinGitEnvLookupErrors e
:}
Left (GitOrEnvLookupEnvLookup (MkEnvLookupError "VAR"))

Since: gitrev-internal-0.1