Development.GitRev.Utils

Description

Utils module.

Since: 2.0

Combining Q actions lazily

newtype QFirst e a

Wrapper for Q over Either with a lazier Semigroup. With this, we can run:

  MkQFirst q1 <> MkQFirst q2

This will only execute q2 if q1 returns Left, unlike Q's normal Semigroup instance.

QFirst also collects all errors in Exceptions.

Since: 2.0

Constructors

MkQFirst

Fields

unQFirst :: Q (Either (Exceptions e) a)

mkQFirst :: Q (Either e a) -> QFirst e a

Since: 2.0

firstSuccessQ :: Q (Either e a) -> [Q (Either e a)] -> Q (Either (Exceptions e) a)

firstSuccessQ q qs takes the first qi in q : qs that returns Right, without executing any qj for j > i. If there are no Right's, returns the final result.

Examples

Expand

>>> :{
   $$( qToCode $
         firstSuccessQ
           (pure (Left GitNotFound))
           [ gitHashQ,
             error "oh no"
           ]
     )
:}
Right ...

Since: 2.0

newtype Exceptions e

Collects multiple exceptions.

Since: 2.0

Constructors

MkExceptions

Fields

unExceptions :: NonEmpty e

mkExceptions :: e -> Exceptions e

Since: 2.0

Either projections

projectStringUnknown :: Functor f => f (Either e String) -> f String

Projects Left to the string UNKNOWN.

Examples

Expand

>>> :{
  let gitHashUnknownQ :: Q String
      gitHashUnknownQ = projectStringUnknown gitHashQ
  -- inling gitHashUnknownQ here due to stage restriction
  in $$(qToCode $ projectStringUnknown gitHashQ)
:}
...

>>> $$(qToCode $ projectStringUnknown (pure $ Left ()))
"UNKNOWN"

Since: 2.0

projectString :: Functor f => String -> f (Either e String) -> f String

Projects Left to the given string.

Examples

Expand

>>> :{
  let gitHashDefStringQ :: Q String
      gitHashDefStringQ = projectString "FAILURE" gitHashQ
  in $$(qToCode $ projectString "FAILURE" gitHashQ)
:}
...

>>> $$(qToCode $ projectString "FAILURE" (pure $ Left ()))
"FAILURE"

Since: 2.0

projectFalse :: Functor f => f (Either e Bool) -> f Bool

Projects Left to False.

Examples

Expand

>>> :{
  let gitDirtyDefFalseQ :: Q Bool
      gitDirtyDefFalseQ = projectFalse gitDirtyQ
  in $$(qToCode $ projectFalse gitDirtyQ)
:}
...

>>> $$(qToCode $ projectFalse (pure $ Left ()))
False

Since: 2.0

projectError :: forall f e a. (Exception e, Functor f) => f (Either e a) -> f a

Projects Left via error, rendering via displayException. Hence an error will cause a compilation failure.

Examples

Expand

>>> :{
  let gitHashOrDieQ :: Q String
      gitHashOrDieQ = projectError gitHashQ
  in $$(qToCode $ projectError gitHashQ)
:}
...

Since: 2.0

projectErrorMap :: Functor f => (e -> String) -> f (Either e a) -> f a

Projects Left via error, rendering via the given function. Hence an error will cause a compilation failure.

Examples

Expand

>>> :{
  let gitHashOrDieQ :: Q String
      gitHashOrDieQ = (projectErrorMap show) gitHashQ
  in $$(qToCode $ (projectErrorMap show) gitHashQ)
:}
...

Since: 2.0

Composing errors

data GitOrLookupEnvError

Git or env lookup error.

Since: 2.0

Constructors

GitOrLookupEnvGit GitError
Since: 2.0

GitOrLookupEnvLookupEnv LookupEnvError
Since: 2.0

Functions

runGitInEnvDirQ

Arguments

:: String
Environment variable pointing to a directory path, in which we run the git process.

-> Q (Either GitError a)
Git process to run.

-> Q (Either GitOrLookupEnvError a)
The result.

runGitInEnvDirQ var q runs q in the directory given by the environment variable.

Examples

Expand

>>> setEnv "SOME_DIR" "./"
>>> $$(qToCode $ runGitInEnvDirQ "SOME_DIR" gitHashQ)
Right ...

Since: 2.0

Mapping utilities

embedGitError

Arguments

:: forall f p a. (Bifunctor p, Functor f)

=> f (p GitError a)
.

-> f (p GitOrLookupEnvError a)

Embeds a GitError in the larger GitOrLookupEnvError.

Examples

Expand

>>> :{
  let q :: Q (Either GitError ())
      q = pure (Left GitNotFound)
  in runQ $ embedGitError q
:}
Left (GitOrLookupEnvGit GitNotFound)

Since: 2.0

embedLookupEnvError

Arguments

:: forall f p a. (Bifunctor p, Functor f)

=> f (p LookupEnvError a)
.

-> f (p GitOrLookupEnvError a)

Embeds a LookupEnvError in the larger GitOrLookupEnvError.

Examples

Expand

>>> :{
  let q :: Q (Either LookupEnvError ())
      q = pure (Left $ MkLookupEnvError "VAR")
  in runQ $ embedLookupEnvError q
:}
Left (GitOrLookupEnvLookupEnv (MkLookupEnvError "VAR"))

Since: 2.0

joinLookupEnvGitErrors

Arguments

:: (Bifunctor p, forall e. Monad (p e))

=> p LookupEnvError (p GitError a)
.

-> p GitOrLookupEnvError a

Utility function for joining lookup and git errors.

Examples

Expand

>>> :{
  let e :: Either LookupEnvError (Either GitError ())
      e = Right (Left GitNotFound)
  in joinLookupEnvGitErrors e
:}
Left (GitOrLookupEnvGit GitNotFound)

Since: 2.0

joinGitLookupEnvErrors

Arguments

:: (Bifunctor p, forall e. Monad (p e))

=> p GitError (p LookupEnvError a)
.

-> p GitOrLookupEnvError a

Utility function for joining git and lookup errors.

Examples

Expand

>>> :{
  let e :: Either GitError (Either LookupEnvError ())
      e = Right (Left $ MkLookupEnvError "VAR")
  in joinGitLookupEnvErrors e
:}
Left (GitOrLookupEnvLookupEnv (MkLookupEnvError "VAR"))

Since: 2.0